calculate the product of two numbers

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -9,7 +9,7 @@
 
 int _mul(int a, int b)
 {
-long int prod;
+int prod;
 prod = a * b;
 return (prod);
 }
